requirements.txt
passlib[bcrypt]
python-jose[cryptography]
python-multipart
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
pydantic-core
httpx
python-dotenv
psycopg2-binary 